refactor: 요청 객체 파싱 기능 분리
